+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on Keep a Changelog (https://keepachangelog.com), and this p
 ### Fixed
 - Add missing unit test for `Abs` in `EvalExpr`.
 - Fix error in unit test for `Abs` in `EvalExpr`.
+- Add missing unit test for `Sgn` in `EvalExpr`.
 
 ## [Unreleased] - 2026-02-14
 ### Changed